Extracted an abstract test class for the network tests.
